// src/app/api/model/server.ts
/**
 * Broker.ManagementApi
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface Server { 
    id?: number;
    serverIp: string | null;
    tenant: string | null;
    username: string | null;
    domain: string | null;
    cF_ZONE_ID: string | null;
    cF_API_TOKEN: string | null;
    giT_URL: string | null;
    branch: string | null;
    giT_TOKEN: string | null;
    password?: string | null;
    log?: string | null;
    processStatus: ServerProcessStatusEnum;
    createdAt: string;
    updatedAt?: string | null;
    url: string | null;
    type: ServerTypeEnum;
    freeUserCount?: number;
    priceByUser?: number;
    balance?: number;
    totalUserCount?: number;
    totalSizeGB?: number | null;
    usedSizeGB?: number | null;
}
export enum ServerProcessStatusEnum {
    CreationInQueue = 'CreationInQueue',
    Creating = 'Creating',
    Running = 'Running',
    DeletionInQueue = 'DeletionInQueue',
    Deleting = 'Deleting',
    PausingInQueue = 'PausingInQueue',
    Pausing = 'Pausing',
    Paused = 'Paused',
    ResumingInQueue = 'ResumingInQueue',
    Resuming = 'Resuming',
    Error = 'Error'
};
export enum ServerTypeEnum {
    AdminApi = 'AdminApi',
    AdminWeb = 'AdminWeb',
    BrokerApi = 'BrokerApi',
    BrokerWeb = 'BrokerWeb'
};