// src/app/api/model/createServerRequest.ts
/**
 * Broker.ManagementApi
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CreateServerRequest { 
    serverIp?: string | null;
    tenant?: string | null;
    username?: string | null;
    domain?: string | null;
    cF_ZONE_ID?: string | null;
    cF_API_TOKEN?: string | null;
    giT_URL?: string | null;
    branch?: string | null;
    giT_TOKEN?: string | null;
    password?: string | null;
}