// src/app/api/model/pausePlayServerRequest.ts
/**
 * Broker.ManagementApi
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface PausePlayServerRequest { 
    serverIp?: string | null;
    tenant?: string | null;
    username?: string | null;
    password?: string | null;
    command?: string | null;
    isApi?: boolean;
}